refactor(#26): 저장위치 변경 및 의존성 변경
