+++ src/main/java/savepay/savepay/global/security/domain/token/repository/RefreshTokenRepository.java
@@ -1,8 +1,8 @@
-package savepay.savepay.global.security.domain.repository;
+package savepay.savepay.global.security.domain.token.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import savepay.savepay.domain.user.entity.User;
-import savepay.savepay.global.security.domain.RefreshToken;
+import savepay.savepay.global.security.domain.token.entity.RefreshToken;
 
 import java.util.Optional;
