+++ src/main/java/savepay/savepay/global/security/domain/token/service/TokenService.java
@@ -1,4 +1,4 @@
-package savepay.savepay.global.security.domain.service;
+package savepay.savepay.global.security.domain.token.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -11,8 +11,8 @@ import savepay.savepay.domain.user.repository.UserRepository;
 import savepay.savepay.global.code.status.ErrorStatus;
 import savepay.savepay.global.exception.GeneralException;
 import savepay.savepay.global.security.JwtTokenProvider;
-import savepay.savepay.global.security.domain.RefreshToken;
-import savepay.savepay.global.security.domain.repository.RefreshTokenRepository;
+import savepay.savepay.global.security.domain.token.entity.RefreshToken;
+import savepay.savepay.global.security.domain.token.repository.RefreshTokenRepository;
 import savepay.savepay.global.security.service.CustomUserDetailsService;
 
 import java.util.Optional;
