+++ src/main/java/savepay/savepay/global/security/filter/CustomJwtFilter.java
@@ -9,7 +9,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import savepay.savepay.global.security.domain.service.TokenService;
+import savepay.savepay.global.security.domain.token.service.TokenService;
 
 import java.io.IOException;
